+++ src/NeverNull/Combinators/ToNullableExt.cs
@@ -1,9 +1,15 @@
-﻿using JetBrains.Annotations;
+﻿using System;
+using JetBrains.Annotations;
 
 namespace NeverNull.Combinators {
+    /// <summary>
+    ///     Provides extension methods to transform instances of <see cref="Option{T}" /> into their <see cref="Nullable{T}" />
+    ///     representation.
+    /// </summary>
     public static class ToNullableExt {
         /// <summary>
-        ///     Returns a nullable containing the value of this option or an empty nullable otherwise.
+        ///     Returns a <see cref="Nullable{T}" /> containing the value of the specified <see cref="Option{T}" /> or an empty
+        ///     <see cref="Nullable{T}" /> otherwise.
         /// </summary>
         /// <typeparam name="T"></typeparam>
         /// <param name="option"></param>