+++ tests/NeverNull.Tests/OptionTests.cs
@@ -28,5 +28,26 @@ namespace NeverNull.Tests {
                 return option.CompareTo(Option.From(x), Comparer.Default) == 0
                     && option.Equals(Option.From(x));
             });
+
+        [Test]
+        public void IfSome_should_only_be_called_on_options_containing_a_value() =>
+            Prop.ForAll<string>(x => {
+                var modifiedVal = "-1";
+                Option.From(x).IfSome(v => modifiedVal = v + "1");
+
+                return modifiedVal.Equals(x == null ? "-1" : x + "1");
+            }).QuickCheckThrowOnFailure();
+
+        [TestFixture]
+        class IfNoneExtTests {
+            [Test]
+            public void IfNone_should_only_be_called_on_None() =>
+                Prop.ForAll<string>(x => {
+                    var modfiedVal = "-1";
+                    Option.From(x).IfNone(() => modfiedVal = "1");
+
+                    return modfiedVal.Equals(x == null ? "1" : "-1");
+                }).QuickCheckThrowOnFailure();
+        }
     }
 }